# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Organization:
    properties:
      created_at:
        type: string
      created_by:
        description: user_id
        type: string
      description:
        type: string
      display_name:
        description: 显示名称
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        description: 唯一标识名称
        type: string
      settings:
        additionalProperties: true
        description: 组织配置
        type: object
      updated_at:
        type: string
    type: object
...
  entity.Team:
    properties:
      created_at:
        type: string
      created_by:
        description: 创建人user_id
        type: string
      description:
        type: string
      display_name:
        description: 显示名称
        type: string
      id:
        type: string
      is_system:
        description: 是否为系统预置团队（不可删除）
        type: boolean
      members:
        items:
          $ref: '#/definitions/entity.TeamMember'
        type: array
      name:
        description: 团队名称
        type: string
      org_id:
        description: 所属组织ID
        type: integer
      organization:
        allOf:
        - $ref: '#/definitions/entity.Organization'
        description: 关联
      updated_at:
        type: string
    type: object
  entity.TeamMember:
    properties:
      id:
        type: integer
      joined_at:
        description: 加入时间
        type: string
      joined_by:
        description: 添加人user_id
        type: string
      role:
        allOf:
        - $ref: '#/definitions/entity.TeamRole'
        description: 团队内角色
      team:
        allOf:
        - $ref: '#/definitions/entity.Team'
        description: 关联
      team_id:
        type: string
      user_id:
        type: string
    type: object
  entity.TeamRole:
    enum:
    - MEMBER
    - MAINTAINER
    type: string
    x-enum-varnames:
    - TeamRoleMember
    - TeamRoleMaintainer
...
  handlers.AddTeamMemberRequest:
    properties:
      role:
        description: MEMBER/MAINTAINER
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
...
  handlers.CreateTeamRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      name:
        type: string
      org_id:
        type: integer
    required:
    - name
    - org_id
    type: object
  handlers.CreateTeamTokenRequest:
    properties:
      expires_in_days:
        description: 0表示永不过期
        type: integer
      token_name:
        type: string
    required:
    - token_name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@iac-platform.com
    name: API Support
  description: Infrastructure as Code Platform API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAC Platform API
  version: "1.0"
paths:
...
  /api/v1/iam/teams/{id}/tokens:
    get:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出团队Token
      tags:
      - IAM-Team
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      - description: 创建Token请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTeamTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: 创建团队Token
      tags:
      - IAM-Team
  /api/v1/iam/teams/{id}/tokens/{token_id}:
    delete:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token ID
        in: path
        name: token_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 吊销团队Token
      tags:
      - IAM-Team
...
  /api/v1/teams:
    get:
      parameters:
      - description: 组织ID
        in: query
        name: org_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Team'
            type: array
      summary: 列出组织的所有团队
      tags:
      - Team
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建团队请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Team'
      summary: 创建团队
      tags:
      - Team
  /api/v1/teams/{id}:
    delete:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除团队
      tags:
      - Team
    get:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Team'
      summary: 获取团队详情
      tags:
      - Team
  /api/v1/teams/{id}/members:
    get:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.TeamMember'
            type: array
      summary: 列出团队成员
      tags:
      - Team
    post:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      - description: 添加成员请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddTeamMemberRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 添加团队成员
      tags:
      - Team
  /api/v1/teams/{id}/members/{user_id}:
    delete:
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 移除团队成员
      tags:
      - Team
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"